+++ packages/evm-module/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @avalabs/evm-module
 
+## 0.11.14
+
+### Patch Changes
+
+- @avalabs/vm-module-types@0.11.14
+
+## 0.11.13
+
+### Patch Changes
+
+- @avalabs/vm-module-types@0.11.13
+
 ## 0.11.12
 
 ### Patch Changes